Fix population_script

+++ population_script.py
@@ -13,11 +13,14 @@ from django.db.models import Avg
 
 def clear():
     superuser = User.objects.filter(is_superuser=True).first()
-    User.objects.exclude(username=superuser.username).delete()
-    UserProfile.objects.exclude(user=superuser).delete()
-    Album.objects.all().delete()
-    Song.objects.all().delete()
-    SongRating.objects.all().delete()
+    if superuser:
+        User.objects.exclude(username=superuser.username).delete()
+        UserProfile.objects.exclude(user=superuser).delete()
+        Album.objects.all().delete()
+        Song.objects.all().delete()
+        SongRating.objects.all().delete()
+    else:
+        print("No superuser found No action taken")
 
 def populate():
